+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## UNRELEASED
 
 - fix: AssemblyScript ESLint import path [#737](https://github.com/hypermodeinc/modus/pull/737)
+- fix: decode base64 strings to binary data types [#738](https://github.com/hypermodeinc/modus/pull/738)
 
 ## 2025-01-24 - Runtime 0.17.1